pyproject.toml
[project]
name = "carvera-cli"
version = "0.1.0"
description = "A command-line tool for managing Carvera CNC machines"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "gnureadline>=8.2.13",
    "pyquicklz>=1.4.1",
    "pyserial>=3.5",
]

[project.scripts]
carvera-cli = "carvera_cli:cli_main"

[project.urls]
"Homepage" = "https://github.com/Carvera-Community/Carvera_Controller"
"Bug Tracker" = "https://github.com/Carvera-Community/Carvera_Controller/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
